+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
 * Fix `holdDyn` so that it is lazy in its event argument
 * Relax constraint from `Applicative` to `Apply` in `ffor2` & `ffor3`. `Event`s and `Map`s can now use these utils.
+* Generalize `leftmost` and `mergeWith` family of functions from `[]` to any `Foldable`.
 
 ## 0.6.1.0